Expiration Date Handling

+++ modules/registrars/keysystems/keysystems.php
@@ -39,6 +39,7 @@ use WHMCS\Module\Registrar\RRPproxy\Commands\QueryZoneList;
 use WHMCS\Module\Registrar\RRPproxy\Commands\RenewDomain;
 use WHMCS\Module\Registrar\RRPproxy\Commands\ResendNotification;
 use WHMCS\Module\Registrar\RRPproxy\Commands\SetAuthCode;
+use WHMCS\Module\Registrar\RRPproxy\Commands\SetDomainRenewalMode;
 use WHMCS\Module\Registrar\RRPproxy\Commands\StatusContact;
 use WHMCS\Module\Registrar\RRPproxy\Commands\StatusDomain;
 use WHMCS\Module\Registrar\RRPproxy\Commands\StatusDomainTransfer;
@@ -151,6 +152,12 @@ function keysystems_getConfigArray(array $params): array
             "Default" => true,
             "Description" => "Enable ID Protection on TLD pricing sync for compatible TLDs"
         ],
+        'RenewProtection' => [
+            'FriendlyName' => 'Renewal Protection',
+            'Type' => 'yesno',
+            "Default" => false,
+            'Description' => 'Skips renewal when domain expiration date is already ahead of the new due date'
+        ],
         'DeleteMode' => [
             'FriendlyName' => 'Domain deletion mode',
             'Type' => 'dropdown',
@@ -162,6 +169,12 @@ function keysystems_getConfigArray(array $params): array
             'Type' => 'yesno',
             "Default" => false,
             'Description' => 'Enables DNSSEC configuration in the client area'
+        ],
+        'DailyCron' => [
+            'FriendlyName' => 'Daily Cron',
+            'Type' => 'yesno',
+            "Default" => false,
+            'Description' => 'Makes some daily consistency checks and sends an e-mail report'
         ],
         'TestMode' => [
             'Type' => 'yesno',
@@ -194,7 +207,7 @@ function keysystems_GetDomainInformation(array $params): Domain
     $domain->setDomain($domainStatus->domainName);
     $domain->setNameservers($domainStatus->nameServers);
     $domain->setTransferLock($domainStatus->transferLock);
-    $domain->setExpiryDate(Carbon::createFromFormat("Y-m-d H:i:s.u", $domainStatus->expirationDate));
+    $domain->setExpiryDate(Carbon::createFromFormat("Y-m-d H:i:s", $domainStatus->expirationDate));
 
     //check contact status
     try {
@@ -289,8 +302,8 @@ function keysystems_RegisterDomain(array $params): array
     try {
         $register->execute();
         return ['success' => true];
-    } catch (Exception $e) {
-        return ['error' => $e->getMessage()];
+    } catch (Exception $ex) {
+        return ['error' => $ex->getMessage()];
     }
 }
 
@@ -319,16 +332,13 @@ function keysystems_TransferDomain(array $params): array
         }
 
         $transfer = new TransferDomain($params);
-        if ($check->userTransferRequired) {
-            $transfer->api->args["ACTION"] = "USERTRANSFER";
-        }
         $transfer->execute();
         if (!empty($transfer->getErrors())) {
             return ["error" => $transfer->getErrors()];
         }
         return ["success" => true];
-    } catch (Exception $e) {
-        return ["error" => $e->getMessage()];
+    } catch (Exception $ex) {
+        return ["error" => $ex->getMessage()];
     }
 }
 
@@ -542,8 +552,8 @@ function keysystems_CheckAvailability(array $params)
             }
         }
         return $results;
-    } catch (Exception $e) {
-        return ['error' => $e->getMessage()];
+    } catch (Exception $ex) {
+        return ['error' => $ex->getMessage()];
     }
 }
 
@@ -575,8 +585,8 @@ function keysystems_GetDomainSuggestions(array $params)
             }
         }
         return $results;
-    } catch (Exception $e) {
-        return ['error' => $e->getMessage()];
+    } catch (Exception $ex) {
+        return ['error' => $ex->getMessage()];
     }
 }
 
@@ -665,8 +675,8 @@ function keysystems_GetEmailForwarding(array $params): array
     try {
         $emailFwd = new MailFwd($params);
         return $emailFwd->values;
-    } catch (Exception $e) {
-        return ['error' => $e->getMessage()];
+    } catch (Exception $ex) {
+        return ['error' => $ex->getMessage()];
     }
 }
 
@@ -683,8 +693,8 @@ function keysystems_SaveEmailForwarding(array $params): array
         $emailFwd = new MailFwd($params);
         $emailFwd->update();
         return [];
-    } catch (Exception $e) {
-        return ['error' => $e->getMessage()];
+    } catch (Exception $ex) {
+        return ['error' => $ex->getMessage()];
     }
 }
 
@@ -778,9 +788,20 @@ function keysystems_ReleaseDomain(array $params): array
  */
 function keysystems_RequestDelete(array $params): array
 {
+    if ($params["DeleteMode"] == "ImmediateIfPossible") {
+        try {
+            $delete = new DeleteDomain($params);
+            $delete->execute();
+            return ['success' => true];
+        } catch (Exception $ex) {
+            // We fallback to setting renewal mode
+        }
+    }
+
     try {
-        $delete = new DeleteDomain($params);
-        $delete->execute();
+        $renewalMode = new SetDomainRenewalMode($params);
+        $renewalMode->setAutoDelete();
+        $renewalMode->execute();
         return ['success' => true];
     } catch (Exception $ex) {
         return ['error' => $ex->getMessage()];
@@ -862,9 +883,9 @@ function keysystems_Sync(array $params): array
         $domain = new StatusDomain($params);
         //TODO set admin/tech/billing contacts if necessary
         return [
-            'active' => $domain->status == 'active',
-            'expired' => $domain->status == 'expired',
-            'expirydate' => Carbon::createFromFormat('Y-m-d H:i:s.u', $domain->expirationDate)->toDateString()
+            'active' => $domain->isActive,
+            'expired' => $domain->isExpired,
+            'expirydate' => Carbon::createFromFormat('Y-m-d H:i:s', $domain->expirationDate)->toDateString()
         ];
     } catch (Exception $ex) {
         return ['error' => $ex->getMessage()];
@@ -895,6 +916,8 @@ function keysystems_TransferSync(array $params): array
             if ($transfer->hasFailed()) {
                 $values['failed'] = true;
                 $values['reason'] = $transfer->getLog();
+            } else {
+                $values['completed'] = false;
             }
         } catch (Exception $ex) {
             $values['error'] = 'StatusDomainTransfer: ' . $ex->getMessage();
@@ -902,7 +925,7 @@ function keysystems_TransferSync(array $params): array
         return $values;
     }
 
-    $values['expirydate'] = Carbon::createFromFormat('Y-m-d H:i:s.u', $domain->expirationDate)->toDateString();
+    $values['expirydate'] = Carbon::createFromFormat('Y-m-d H:i:s', $domain->expirationDate)->toDateString();
 
     $zoneInfo = ZoneInfo::get($params);
     if (!$zoneInfo->renews_on_transfer) {
@@ -926,7 +949,10 @@ function keysystems_TransferSync(array $params): array
         if (count($diffNameservers) > 0) {
             $i = 1;
             foreach ($orderNameservers as $nameserver) {
-                $modifyDomain->params["ns$i"] = $nameserver;
+                if (!$nameserver) {
+                    continue;
+                }
+                $modifyDomain->params["ns" . $i++] = $nameserver;
             }
             $modifyDomain->setNameServers();
         }
@@ -1078,8 +1104,8 @@ function keysystems_GetTldPricing(array $params)
     try {
         $zoneList = new QueryZoneList($params);
         $zoneList->execute();
-    } catch (Exception $e) {
-        return ["error" => "QueryZoneList - " . $e->getMessage()];
+    } catch (Exception $ex) {
+        return ["error" => "QueryZoneList - " . $ex->getMessage()];
     }
 
     foreach ($zoneList->api->properties["ZONE"] as $id => $zone) {
